Fixing bug in latency calculation
In the previous version, the latency of the first gaze shift was always calculated relative to stimulus onset, while the latency of subsequent gaze shifts was calculate relative to the offsets of previous gaze shifts.

This assumed that the first gaze shift always occured after stimulus onset, which is not always true. As a consequence, the latency of the actually first gaze shift after stimulus onset was sometimes massively overestimated